Added the ability to store the key passphrase in the mix.cfg file instead of having it compiled into the binary. This offers a number of advantages over the compiled-in method and still allows for encrypted remote storage of keys.
In order to take advantage of the new changes, the Install script will need to
be modified. As this is being replaced with a proper autoconf system, we're
defering that until later.

The changes we're committing replace PASSPHRASE with PASS_PHRASE, which is read
from the config. If the compiled-in passphrase, PASSPHRASE, is set, that is
used instead. (This is how Mixmaster will function if you use the Install
script.)