Added support for AES in Type I messages. This should prevent lost mail if a user attempts to send OpenPGP messages encrypted with AES.
Support for AES requires OpenSSL 0.9.7 (soon to be released.) The Install
script needs to be updated to enable AES if 0.9.7 or later is present.

Also, pgp_keylen now returns 0 instead of -1 for unknown symalgo. Output
of pgp_keylen is used in buf_appendrnd and in other functions without
checking, so they might be unhappy with -1. (We should really check for
this.)

Thanks to Disastry and Peter for working on this.