Re-introduced the todo list with Reality.
Basically, if the install process is made friendly and saner, and we've
done real testing on the major platforms, I'm not too interested in holding
up 3.0 any longer. Everything else can be done in the point release.

Should we:
* leave AES support #ifdef'd out if OpenSSL 0.9.7 isn't released?
* support AES if the installed version of OpenSSL is 0.9.7beta3 or greater?
* include the necessary AES source from OpenSSL, as per the original patch?

I'm leaning toward option two, though we should still display the "this
version of OpenSSL is untested" message for 0.9.7. FWIW, 0.9.7beta3 has been
working fine for me on randseed.