Make sure to deliver to the right directory with Maildir deliveries. Since we do not chdir() to the Mixmaster base directory we need to take care of this whenever we do file or directory operations. Fortunatly the function mixfile() helps here. mixfile() was extended to handle paths starting with "~/". '~' is replaced by the HOME environment variable if defined.
Implement MAILIN option: If set Mixmaster reads mails from either a mbox
or a Maildir mail folder when processing its queue (-M or daemon). After
reading the mail folder is emtpied. If MAILIN ends with / it is expected
to be a Maildir, mbox oterhwise.

I hope that I did not break anything with the change to mixfile(). Maildir
drop and reading from both Maildir and mbox has been tested a little. Seems
to work fine so far.