Fixed problem with stats. Numbers at 00:00 GMT were wrong.
The problem was in expiring hourly stats to daily stats: Once a dataset
is older than 24 hours it gets moved to the daily bucket. Unfortunatly
daily buckets are aligned at 00:00GMT each day. So next time stats.log
is read that 00:00GMT sample is within our 24hrs again.

The fix is to mark daily sets as such and to never import them to hourly
statistics.