Previosly when we ran Mixmaster in daemon mode and injected mail either via pop3 or the MAILIN method (reading a Maildir or an mbox), Mixmaster only processed remailer-xxx requests at SENDPOOLTIME intervalls.
It might be in the interest of an operator to have a high SENDPOOLTIME -
like say an hour or two - to provide better security to their user.

However it is also important to answer remailer-xxx requests in a timely
manner.

This conflict of interests can be solved by adding yet another config
option: MAILINTIME.

MAILINTIME specifies the amount of time between reading MAILIN and
processing those mails (as well as those injected via POP3). Processing
here means to answer -xxx requests and to add remailer mails to the
pool.

This is also done implicitly whenever Mixmaster is about to process its
pool or just fetched mail via pop3.