Mixmaster now prompts for user input for the secret key passphrase when mix is started in daemon mode (./mix -D). Mixmaster will try to find either PGP/RSA, PGP/ELG or Mix key that can be decrypted with the provided passphrase. If it finds one, it assumes the passphrase is good.
(We may have a problem if the user manages to get his secret key
passphrases out of sync. Suggestions on handling this case?)

FYI: the passphrase in mix.cfg takes precedence over the passphrase
compiled into the binary. If neither of these are present, only then is
the user prompted for a keyboard-input passphrase.

We still need to make mix retain the user-provided passphrase across
SIGHUP signals.