When creating new OpenPGP keys, also set an expiry date. Key lifetime defaults to 8 months but can be overriden by the KEYLIFETIME configuration option.
We currently do not store the self signature and the keybinding (which hold
the expiry information in DSA keys) in the secret keyring. This is
unfortunate because we use the current KEYLIFETIME when recreating them
should the public keyring need to be rewritten. The solution is to store
them in the secret keyring (like GnuPG does) and not recreate them later
if we already have them.