Do not recreate the public keys from the private keyring if there still is a public keyring in pgpkey.txt. This means we don't recreate the self signature on binding keys amoung other things. It probably also means that one can replace the keys in pgpkey.txt with keys signed by the adminkey and not have them overriden the next time one does ./mix -[GK] (although I did not test this).
We still need to fix keymgt to only return the latest keys - this should be
simple once I find out how to get expiration date for a specific key.

There's a small problem with the current keymanagement for PGP keys I guess. If
the latest key is not the last in the keyring we might screw up. Can you verify
this Disastry?