Basic autoconf support.
This is just a beginning.  Len says it has problems detecting his
screwy OpenSSL installation, which we can work on.  Also, I didn't
start mucking about in using the autoconf-generated config.h instead
of a bunch of -D options on the compiler command line.  This is
something we should do, though!

There is also not yet a replacement for the installation part of the
old 'Install' script.