Changed the COPYRIGHT definition to be more inclusive.
I'd prefer to make this "The Mixmaster Group" and list the copyright holders as a subset of that in the COPYRIGHT file, but I'm not sure I can do that without there being a legal entity bearing that name.

Does one list it as "Copyright Foo" or "(C) Foo"? Or Can we use the actual copyright symbol?